refactor added for middle section
