+++ src/features/sign-in-sign-up/types.ts
@@ -1,0 +1,11 @@
+export type MiddleSectionProps = {
+  googleLabel: string;
+  facebookLabel: string;
+  ctaAnnouncementLabel: string;
+  ctaLabel: string;
+  ctaLink: string;
+};
+
+export type SignInSignUpProps = {
+  middleSection: MiddleSectionProps;
+};
